#35 - split model classes to users and teams packages
