+++ src/main/java/ua/com/juja/microservices/teams/slackbot/model/teams/ActivateTeamRequest.java
@@ -1,5 +1,6 @@
-package ua.com.juja.microservices.teams.slackbot.model;
+package ua.com.juja.microservices.teams.slackbot.model.teams;
 
+import lombok.Getter;
 import lombok.ToString;
 import lombok.extern.slf4j.Slf4j;
 import org.hibernate.validator.constraints.NotEmpty;
@@ -12,12 +13,15 @@ import java.util.Set;
  */
 @ToString
 @Slf4j
-public class TeamRequest {
-
+@Getter
+public class ActivateTeamRequest {
+    @NotEmpty
+    private String from;
     @NotEmpty
     private final Set<String> members;
 
-    public TeamRequest(Set<String> members) {
+    public ActivateTeamRequest(String from, Set<String> members) {
+        this.from = from;
         this.members = members;
     }
