+++ src/main/java/ua/com/juja/microservices/teams/slackbot/model/teams/Team.java
@@ -1,6 +1,5 @@
-package ua.com.juja.microservices.teams.slackbot.model;
+package ua.com.juja.microservices.teams.slackbot.model.teams;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
@@ -17,10 +16,16 @@ import java.util.Set;
 @ToString
 @Slf4j
 public class Team {
+    @JsonProperty("members")
+    private final Set<String> members;
+
+    @JsonProperty("from")
+    private String from;
 
     @Getter
     @JsonProperty("id")
     private String id;
+
     @JsonProperty("activateDate")
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ssZ")
     private Date activateDate;
@@ -29,16 +34,27 @@ public class Team {
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ssZ")
     private Date deactivateDate;
 
-    @JsonProperty("members")
-    private final Set<String> members;
+    public Team(@JsonProperty("from") String from, @JsonProperty("id") String id,
+                @JsonProperty("activateDate") Date activateDate, @JsonProperty("deactivateDate") Date deactivateDate,
+                @JsonProperty("members") Set<String> members) {
+        this.from = from;
+        this.id = id;
+        this.activateDate = activateDate;
+        this.deactivateDate = deactivateDate;
+        this.members = members;
+    }
 
-    @JsonCreator
     public Team(@JsonProperty("members") Set<String> members) {
+        this.members = members;
+    }
+
+    public Team(@JsonProperty("from") String from,
+                @JsonProperty("members") Set<String> members) {
+        this.from = from;
         this.members = members;
     }
 
     public Set<String> getMembers() {
         return Collections.unmodifiableSet(members);
     }
-
 }
