+++ src/main/java/ua/com/juja/microservices/teams/slackbot/model/users/User.java
@@ -1,4 +1,4 @@
-package ua.com.juja.microservices.teams.slackbot.model;
+package ua.com.juja.microservices.teams.slackbot.model.users;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
