+++ src/main/java/ua/com/juja/microservices/teams/slackbot/model/users/UserSlackNameRequest.java
@@ -1,6 +1,7 @@
-package ua.com.juja.microservices.teams.slackbot.model;
+package ua.com.juja.microservices.teams.slackbot.model.users;
 
 import lombok.AllArgsConstructor;
+import lombok.ToString;
 
 import java.util.Collections;
 import java.util.List;
@@ -9,6 +10,7 @@ import java.util.List;
  * @author Ivan Shapovalov
  */
 @AllArgsConstructor
+@ToString
 public class UserSlackNameRequest {
     List<String> slackNames;
