+++ src/main/java/ua/com/juja/microservices/teams/slackbot/repository/TeamRepository.java
@@ -1,16 +1,17 @@
 package ua.com.juja.microservices.teams.slackbot.repository;
 
-import ua.com.juja.microservices.teams.slackbot.model.Team;
-import ua.com.juja.microservices.teams.slackbot.model.TeamRequest;
+import ua.com.juja.microservices.teams.slackbot.model.teams.ActivateTeamRequest;
+import ua.com.juja.microservices.teams.slackbot.model.teams.DeactivateTeamRequest;
+import ua.com.juja.microservices.teams.slackbot.model.teams.Team;
 
 /**
  * @author Ivan Shapovalov
  */
 public interface TeamRepository {
 
-    Team activateTeam(TeamRequest teamRequest);
+    Team activateTeam(ActivateTeamRequest activateTeamRequest);
 
-    Team deactivateTeam(String slackName);
+    Team deactivateTeam(DeactivateTeamRequest deactivateTeamRequest);
 
     Team getTeam(String slackName);
