+++ src/main/java/ua/com/juja/microservices/teams/slackbot/repository/impl/RestUserRepository.java
@@ -11,9 +11,9 @@ import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.client.RestTemplate;
 import ua.com.juja.microservices.teams.slackbot.exceptions.ApiError;
 import ua.com.juja.microservices.teams.slackbot.exceptions.UserExchangeException;
-import ua.com.juja.microservices.teams.slackbot.model.User;
-import ua.com.juja.microservices.teams.slackbot.model.UserSlackNameRequest;
-import ua.com.juja.microservices.teams.slackbot.model.UserUuidRequest;
+import ua.com.juja.microservices.teams.slackbot.model.users.User;
+import ua.com.juja.microservices.teams.slackbot.model.users.UserSlackNameRequest;
+import ua.com.juja.microservices.teams.slackbot.model.users.UserUuidRequest;
 import ua.com.juja.microservices.teams.slackbot.repository.UserRepository;
 import ua.com.juja.microservices.teams.slackbot.util.SlackNameHandler;
 import ua.com.juja.microservices.teams.slackbot.util.Utils;
